Проверьте, пожалуйста, дз Java Collections Framework: Set Вынес SearchableComparator в отдельный файл
